move buttons to display page,
